fix: hide approval-gated tools in CLI chat
