+++ packages/cli/src/commands/_helpers/approvalPolicyInstruction.ts
@@ -1,21 +1,14 @@
-export type ApprovalInstructionContext = {
-  readonly approvalPolicy: 'ask' | 'never' | 'yolo';
-  readonly mode: 'headless' | 'interactive';
-};
+import {
+  approvalPromptsUnavailable,
+  type ApprovalInstructionContext,
+} from '@cli/runtime/approvalPolicyAvailability';
+
+export { approvalPromptsUnavailable, type ApprovalInstructionContext };
 
 const PRIVILEGED_ACTION_GUIDANCE =
   'Do not call approval-gated tools such as bash/shell commands, file edits, setup/config updates, user questions, retry/plan approvals, or new subagent delegations; solve from the provided context or state what approval is needed.';
 const CLI_APPROVAL_POLICY_GUIDANCE =
   'Valid CLI approval policies are "ask", "never", and "yolo" only. If suggesting a rerun, use --approval-policy yolo for headless auto-approval or an interactive run with --approval-policy ask; do not invent other approval mode names.';
-
-export function approvalPromptsUnavailable(
-  context: ApprovalInstructionContext,
-): boolean {
-  return (
-    context.approvalPolicy === 'never' ||
-    (context.mode === 'headless' && context.approvalPolicy === 'ask')
-  );
-}
 
 export function formatUnavailableApprovalInstruction(
   context: ApprovalInstructionContext,
